Addressed some bugs related to where Firefox places the cursor when it is between the end of a non-formatted text node and the start of a formatted text node.
Added handling for input of non-word characters in the 'input'
event to ensure that text that requires formatting is actually
formatted. Before that, the non-word characters were mainly
handled in the 'beforeinput' event which would only remove
formatting when needed. This resulted in issues where if you
have a word such as '#hashtag1#hashtag2' and the user adds
a non-word character, say a space, before the second hash
character, making the text '#hashtag1 #hashtag2', the two
newly separated hashtags won't both be formatted, which they
should be.