Replaced the findFirstSpaceInText and findLastSpaceInText functions with findFirstNonwordInText and findLastNonwordInText. The two new functions look for any non-word characters that aren't the # character. This is because what Twitter treats as a word is any series of word characters bounded by non-word characters, not just by spaces.
The result of this change is that if the user types something such
as '/#hello', the getCurrentWord function will return '#hello'
instead of '/#hello'. The former will match the hashtag pattern, so
it will be highlighted. The latter, on the other hand, won't match
the pattern and won't be highlighted. Twitter works in the former
way, so the functions had to be replaced to match Twitter's
behavior.